+++ src/el-data-table.vue
@@ -197,7 +197,7 @@ import _get from 'lodash.get'
 import SelfLoadingButton from './self-loading-button.vue'
 import TextButton from './text-button.vue'
 import * as queryUtil from './utils/query'
-import getSelectStrategy from './utils/select-strategies'
+import getSelectStrategy from './utils/select-strategy'
 
 // 默认返回的数据格式如下
 //          {
